Fixed the Spaghetti
Added FaceRecogtinionController that detects faces and does all the stuff regarding it.
Drawing is inside this class too.
Before the loop, create a new instance.
Call Initialize.
In the Loop call the DoLoop method with appropriate parameters.
After the Loop is over and we want to end our work, we have to save the data collected with SaveData() method.